NEW: "binary_debian-installer-includes" config directory.
This add another "-includes" power to live-helper, just leave your full
path files there and they will end on the installer initrd. This is
mostly useful for hacking the installer images with themes, firmware and
new installation helpers.

"This doesn't handle the case where you want to remove files or
 programmatically modify things" yet.

Signed-off-by: Marco Amadori <amadorim@vdavda.com>